Title: 학점 계산기, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 231230/학점 계산기/credit-calculator.c
@@ -10,7 +10,7 @@ int main() {
         scanf("%f",&a[i]);
         sum+=a[i];
     }
-    printf("%.1f\n",(float)sum/n);
+    printf("%.1f\n",sum/n);
     if((float)sum/n>=4){
         printf("Perfect");
